migrate FdSet, nodelay getsockopt, accept rawfd
