+++ src/interface/types.rs
@@ -1,8 +1,6 @@
 #![allow(dead_code)]
 use crate::interface;
 use crate::interface::errnos::{Errno, syscall_error};
-use std::os::unix::io::{AsRawFd, RawFd};
-use std::mem;
 
 const SIZEOF_SOCKADDR: u32 = 16;
 
@@ -274,10 +272,10 @@ pub fn get_mutcbuf_null(union_argument: Arg) -> Result<Option<*mut u8>, i32> {
     return Ok(None);
 }
 
-pub fn get_fdset(union_argument: Arg) -> Result<Option<&'static mut FdSet>, i32> {
+pub fn get_fdset(union_argument: Arg) -> Result<Option<&'static mut interface::FdSet>, i32> {
     let data: *mut libc::fd_set = unsafe{union_argument.dispatch_fdset};
     if !data.is_null() {
-        let internal_fds: &mut FdSet = FdSet::new_from_ptr(data);
+        let internal_fds: &mut interface::FdSet = interface::FdSet::new_from_ptr(data);
         return Ok(Some(internal_fds));
     }
     return Ok(None);
@@ -422,107 +420,6 @@ pub fn get_sockpair<'a>(union_argument: Arg) -> Result<&'a mut SockPair, i32> {
         return Ok(unsafe{&mut *pointer});
     }
     return Err(syscall_error(Errno::EFAULT, "dispatcher", "input data not valid"));
-}
-
-pub struct FdSet(libc::fd_set);
-
-impl FdSet {
-    pub fn new() -> FdSet {
-        unsafe {
-            let mut raw_fd_set = std::mem::MaybeUninit::<libc::fd_set>::uninit();
-            libc::FD_ZERO(raw_fd_set.as_mut_ptr());
-            FdSet(raw_fd_set.assume_init())
-        }
-    }
-
-    pub fn new_from_ptr(raw_fdset_ptr: *const libc::fd_set) -> &'static mut FdSet {
-        unsafe {
-            &mut *(raw_fdset_ptr as *mut FdSet)
-        }
-    }
-
-    // copy the src FdSet into self
-    pub fn copy_from(&mut self, src_fds: &FdSet) {
-        unsafe {
-            std::ptr::copy_nonoverlapping(&src_fds.0 as *const libc::fd_set, &mut self.0 as *mut libc::fd_set, 1);
-        }
-    }
-
-    // turn off the fd bit in fd_set
-    pub fn clear(&mut self, fd: RawFd) {
-        unsafe { libc::FD_CLR(fd, &mut self.0) }
-    }
-
-    // turn on the fd bit in fd_set
-    pub fn set(&mut self, fd: RawFd) {
-        unsafe { libc::FD_SET(fd, &mut self.0) }
-    }
-
-    // return true if the bit for fd is set, false otherwise
-    pub fn is_set(&self, fd: RawFd) -> bool {
-        unsafe { libc::FD_ISSET(fd, &self.0) }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        let fd_array: &[u8] = unsafe {
-            std::slice::from_raw_parts(&self.0 as *const _ as *const u8, mem::size_of::<libc::fd_set>())
-        };
-        fd_array.iter().all(|&byte| byte == 0)
-    }
-
-    // for each fd, if kernel_fds turned it on, then self will turn the corresponding tranlated fd on
-    pub fn set_from_kernelfds_and_translate(&mut self, kernel_fds: &FdSet, nfds: i32, rawfd_lindfd_tuples: &Vec<(i32, i32)>) {
-        for fd in 0..nfds {
-            if kernel_fds.is_set(fd) {
-                for (rawfd, lindfd) in rawfd_lindfd_tuples {
-                    if *rawfd == fd {
-                        self.set(*lindfd);
-                    }
-                }
-            }
-        }
-    }
-}
-
-// for unwrapping in kernel_select
-fn to_fdset_ptr(opt: Option<&mut FdSet>) -> *mut libc::fd_set {
-    match opt {
-        None => std::ptr::null_mut(),
-        Some(&mut FdSet(ref mut raw_fd_set)) => raw_fd_set,
-    }
-}
-
-pub fn kernel_select(nfds: libc::c_int, readfds: Option<&mut FdSet>, writefds: Option<&mut FdSet>, errorfds: Option<&mut FdSet>, timeout: Option<interface::RustDuration>) -> i32 {
-    // Call libc::select and store the result
-    let result = unsafe {
-        // Create a timeval struct with zero timeout
-
-        // let mut kselect_timeout = match timeout {
-        //     Some(duration) => libc::timeval {
-        //             tv_sec: duration.as_secs() as i64,
-        //             tv_usec: duration.subsec_micros() as i64,
-        //     },
-        //     None => libc::timeval {
-        //         tv_sec: interface::RustDuration::from_secs(1).as_secs() as i64,  // 0 seconds
-        //         tv_usec: 0, // 0 microseconds
-        //     }
-        // };
-
-        let mut kselect_timeout = libc::timeval {
-                tv_sec: interface::RustDuration::from_secs(30).as_secs() as i64,  // 0 seconds
-                tv_usec: 0, // 0 microseconds
-            };
-
-        libc::select(
-            nfds,
-            to_fdset_ptr(readfds),
-            to_fdset_ptr(writefds),
-            to_fdset_ptr(errorfds),
-            &mut kselect_timeout as *mut libc::timeval,
-        )
-    };
-
-    return result;
 }
 
 pub fn get_sockaddr(union_argument: Arg, addrlen: u32) -> Result<interface::GenSockaddr, i32> {
